added test for authorization
