+++ api/auth/handler.ts
@@ -4,20 +4,22 @@ const AUTH0_CLIENT_SECRET = 'wvSHEEB3V_VvnuwxIDWSqukWoI3tTcqf28YYpKndZEXn3pYj3Q0
 
 // Policy helper function
 function generatePolicy(principalId, effect, resource) {
-    const authResponse: any = {};
-    authResponse.principalId = principalId;
-    if (effect && resource) {
-        const policyDocument: any = {};
-        policyDocument.Version = '2012-10-17';
-        policyDocument.Statement = [];
-        const statementOne: any = {};
-        statementOne.Action = 'execute-api:Invoke';
-        statementOne.Effect = effect;
-        statementOne.Resource = resource;
-        policyDocument.Statement[0] = statementOne;
-        authResponse.policyDocument = policyDocument;
+    if (!effect || !resource) {
+        return {
+            principalId: principalId
+        };
     }
-    return authResponse;
+    return {
+        principalId: principalId,
+        policyDocument: {
+            Version: '2012-10-17',
+            Statement: [{
+                Action: 'execute-api:Invoke',
+                Effect: effect,
+                Resource: resource
+            }]
+        }
+    };
 }
 
 // Reusable Authorizer function, set on `authorizer` field in serverless.yml
